refactor: rewrite functions as modules using index.js for utils
