+++ src/commands/file/compress.js
@@ -3,8 +3,7 @@ import { access } from 'fs/promises';
 import { join, isAbsolute, basename } from 'path';
 import { createBrotliCompress } from 'zlib';
 import * as directoryCommands from '../directory/index.js';
-import { directoryCheck } from '../../utils/directory-check.js';
-import { fileCheck } from '../../utils/file-check.js';
+import * as utils from '../../utils/index.js';
 
 export const compressFile = async (filePath, destinationPath) => {
     try {
@@ -19,8 +18,8 @@ export const compressFile = async (filePath, destinationPath) => {
         await access(fullFilePath);
         await access(fullDestinationPath);
 
-        const isFile = await fileCheck(fullFilePath);
-        const isDirectory = await directoryCheck(fullDestinationPath);
+        const isFile = await utils.fileCheck(fullFilePath);
+        const isDirectory = await utils.directoryCheck(fullDestinationPath);
 
         if (!isFile || !isDirectory) {
             throw new Error('Operation failed');
