+++ src/commands/file/delete.js
@@ -1,7 +1,7 @@
 import { access, unlink } from 'fs/promises';
 import { join, isAbsolute } from 'path';
 import * as directoryCommands from '../directory/index.js';
-import { fileCheck } from '../../utils/file-check.js';
+import * as utils from '../../utils/index.js';
 
 export const deleteFile = async (filePath) => {
     try {
@@ -13,7 +13,7 @@ export const deleteFile = async (filePath) => {
         const fullFilePath = isAbsolute(filePath) ? filePath : join(currentDirectory, filePath);
 
         await access(fullFilePath);
-        const isFile = await fileCheck(fullFilePath);
+        const isFile = await utils.fileCheck(fullFilePath);
         if (!isFile) {
             throw new Error('Operation failed');
         }
