+++ src/commands/file/hash.js
@@ -3,7 +3,7 @@ import { createReadStream } from 'fs';
 import { join, isAbsolute } from 'path';
 import { createHash } from 'crypto';
 import * as directoryCommands from '../directory/index.js';
-import { fileCheck } from '../../utils/file-check.js';
+import * as utils from '../../utils/index.js';
 
 export const calculateHash = async (filePath) => {
     try {
@@ -14,7 +14,7 @@ export const calculateHash = async (filePath) => {
         const fullFilePath = isAbsolute(filePath) ? filePath : join(directoryCommands.getCurrentDirectory(), filePath);
         await access(fullFilePath);
 
-        const isFile = await fileCheck(fullFilePath);
+        const isFile = await utils.fileCheck(fullFilePath);
         if (!isFile) {
             throw new Error('Operation failed');
         }
