+++ src/utils/check-for-quotes.js
@@ -1,0 +1,3 @@
+export const checkForQuotes = (str) => {
+    return /['"]/.test(str);
+};
