+++ src/utils/close-process.js
@@ -1,0 +1,4 @@
+export const closeProcess = (name) => {
+    console.log(`Thank you for using File Manager, ${name}, goodbye!`);
+    process.exit();
+};
