+++ src/utils/index.js
@@ -1,0 +1,15 @@
+import { checkForQuotes } from './check-for-quotes.js';
+import { closeProcess } from './close-process.js';
+import { directoryCheck } from './directory-check.js';
+import { getArgs } from './get-args.js';
+import { fileCheck } from './file-check.js';
+import { splitCommand } from './split-command.js';
+
+export {
+    checkForQuotes,
+    closeProcess,
+    directoryCheck,
+    fileCheck,
+    splitCommand,
+    getArgs,
+  };
